added local setup readme

+++ gmaps/place.go
@@ -173,7 +173,7 @@ func (j *PlaceJob) BrowserActions(ctx context.Context, page playwright.Page) scr
 
 	pageResponse, err := page.Goto(j.GetURL(), playwright.PageGotoOptions{
 		WaitUntil: playwright.WaitUntilStateNetworkidle,
-		Timeout:   playwright.Float(45000), // Increased timeout to 45 seconds for redirects
+		Timeout:   playwright.Float(90000), // Increased timeout to 90 seconds for redirects (with proxies)
 	})
 	if err != nil {
 		resp.Error = err
